Move file and adapt import
