+++ variations/jec.py
